address lucas feedback
